Removed blank spaces from Saved File and implemented export/import modules
