+++ Quantity_Surveying/column.html
@@ -7,7 +7,7 @@
 
     <link rel="stylesheet" href="/styles.css">
     <script src="/layout_script.js" defer></script>
-    <script src="/scriptColumn.js" defer></script>
+    <script src="/scriptColumn.js" defer type="module"></script>
     
 </head>
 <body>
